+++ src/main/java/br/edu/ifsp/spo/bike_integration/repository/PessoaJuridicaRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.edu.ifsp.spo.bike_integration.model.PessoaJuridica;
 
-public interface PessoaJuridicaRepository extends JpaRepository<PessoaJuridica, Integer>{
+public interface PessoaJuridicaRepository extends JpaRepository<PessoaJuridica, Long>{
 
 	Optional<PessoaJuridica> findByCnpj(String cnpj);
 }
